+++ Windows/TrainingWindow.cs
@@ -80,6 +80,13 @@ public sealed class TrainingWindow : Window
                 ImGui.EndTabItem();
             }
 
+            if (ImGui.BeginTabItem("Skill Level"))
+            {
+                ImGui.Spacing();
+                SkillProgressTab.Draw(this.trainingService, this.configuration.Training);
+                ImGui.EndTabItem();
+            }
+
             ImGui.EndTabBar();
         }
     }